requirements.txt
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.23.0
qdrant-client>=1.9.0
